Support for f-string interpolation, Please syntax.
The project inherited its handling of f-strings from the Python
tree-sitter project.  However Starlark doesn't support f-strings, and
the Please build language, which is derived from Starlark, supports only
a simplified and slightly different version of f-strings. This commit
implements the version of f-string interpolation that Please supports.
This differs from Python in the following ways:

1. Python allows arbitrary expressions inside f-string interpolations,
   whereas Please only allows simple variable references.
2. Please adds an additional way of escaping interpolations.  In Python,
   you can escape the `{}` in an f-string by doubling them up.  So
   `f"{{foo}}"` evaluates to `"{foo}"`.  Please supports this, but in
   addition, if you prefix the opening "{" with a '$' it also escapes
   it.  So `f"${foo}"` evalues to '"${foo}"'.  This is done for
   convenience, because it's so common for strings to contain shell
   scripts, which often use the `${varname}` syntax to refer to
   variables.